RD-4629: Extracted component for the modal form field
