+++ app/components/ContactDetailsModal/ContactDetailsModalContent/ContactDetailsModalContent.tsx
@@ -2,9 +2,9 @@ import type { FunctionComponent } from 'react';
 import React, { useMemo } from 'react';
 import { useBoolean, useErrors, useInputs } from '../../../utils/hooks';
 import { Modal, Form, ApproveButton, ErrorMessage } from '../../basic';
-import type { FormField } from './formFields';
+import type { FormField, FormValues } from './formFields';
 import { FormFieldType, getFormFields } from './formFields';
-import CheckboxLabel from './CheckboxLabel';
+import ContactDetailsFormField from './ContactDetailsFormField';
 import StageUtils from '../../../utils/stageUtils';
 import { removeHtmlTagsFromString } from './utils';
 import useManager from '../../../utils/hooks/useManager';
@@ -12,18 +12,12 @@ import Internal from '../../../utils/Internal';
 
 const t = StageUtils.getT('contactDetailsModal.form');
 
-type FormFieldValue = string | boolean;
-
-interface FormInputs {
-    [inputName: string]: FormFieldValue;
-}
-
 interface ContactDetailsModalContentProps {
     closeModal: () => void;
 }
 
 const ContactDetailsModalContent: FunctionComponent<ContactDetailsModalContentProps> = ({ closeModal }) => {
-    const [formInputs, setFormInputs] = useInputs<FormInputs>({});
+    const [formValues, setFormValues] = useInputs<FormValues>({});
     const { errors, setErrors, clearErrors } = useErrors();
     const [hasSubmittingError, showSubmittingError, hideSubmittingError] = useBoolean();
     const [loading, setLoading, cancelLoading] = useBoolean();
@@ -36,7 +30,7 @@ const ContactDetailsModalContent: FunctionComponent<ContactDetailsModalContentPr
     );
 
     const isFieldEmpty = (formField: FormField) => {
-        const fieldValue = formInputs[formField.name];
+        const fieldValue = formValues[formField.name];
         if (formField.type === FormFieldType.Text) {
             return _.isEmpty(fieldValue);
         }
@@ -45,7 +39,7 @@ const ContactDetailsModalContent: FunctionComponent<ContactDetailsModalContentPr
     };
 
     const isFieldValid = (formField: FormField) => {
-        const fieldValue = formInputs[formField.name] as string;
+        const fieldValue = formValues[formField.name] as string;
         const shouldBeValidated = !!formField.validation;
 
         if (shouldBeValidated) {
@@ -83,7 +77,7 @@ const ContactDetailsModalContent: FunctionComponent<ContactDetailsModalContentPr
             setLoading();
             internal
                 .doPost('contactDetails/', {
-                    body: formInputs
+                    body: formValues
                 })
                 .then(() => closeModal())
                 .catch(() => {
@@ -105,27 +99,11 @@ const ContactDetailsModalContent: FunctionComponent<ContactDetailsModalContentPr
                 )}
                 <Form errors={errors} onErrorsDismiss={clearErrors}>
                     {Object.values(formFields).map(formField => (
-                        <Form.Field key={formField.name} required={formField.isRequired}>
-                            {/* TODO: Extract logic a separate component */}
-                            {formField.type === FormFieldType.Text ? (
-                                <Form.Input
-                                    type="text"
-                                    name={formField.name}
-                                    label={formField.label}
-                                    value={formInputs[formField.name]}
-                                    onChange={setFormInputs}
-                                    required={formField.isRequired}
-                                />
-                            ) : (
-                                <Form.Checkbox
-                                    name={formField.name}
-                                    label={<CheckboxLabel label={formField.label} />}
-                                    help=""
-                                    checked={!!formInputs[formField.name]}
-                                    onChange={setFormInputs}
-                                />
-                            )}
-                        </Form.Field>
+                        <ContactDetailsFormField
+                            formField={formField}
+                            formValues={formValues}
+                            onChange={setFormValues}
+                        />
                     ))}
                 </Form>
             </Modal.Content>
