+++ backend/routes/ContactDetails.ts
@@ -43,6 +43,10 @@ router.use(express.json());
 router.get('/', async (req, res) => {
     const token = getTokenFromCookies(req);
 
+    return res.send({
+        contactDetailsReceived: false
+    });
+
     try {
         await checkContactDetailsExistance(token);
         res.send({
